tweeked the assembler hadoop app to read in regular fasta files.
Also:
  - assemblerHadoopApp now reads does most of the work in the reduce step
  - fixed the cleanup of the cap and velvet commands

the consequence of doing the work in the reduce step is that 1) all reads in a
particular group no longer need to be together.  the map task pushes all
reads in a group together.

and you need to set the number of reduce steps to more than the default 1.

see ref 2 and ref 6