dereplicate: reworked the map/reduce steps to reduce the size of the groups.
instead of pushing out each readnodeset, use the hashcode to hash the set of
nodes in the group.  specifically, instead of writing out:
  length=2,<readnode1>,<readnode2>, ...
some groups have sizes in the millions, so writing out that record to disk and
reading it back in the next step takes like forever.  so, instead, generate
the string then write out its hashcode instead, as in:
   <length>.<hashcode>

I add length just in case hashcode is for some magically reason the same.  it
is still not perfect, and there could be two groups that because their
cannonical name is the same get combined.

Refs #8