added BioPig.xml to the etc directory that gets added to the Jar file.
still not sure how to do this.  basically the problem is that pig does
not find the BioPig.xml at runtime.  Setting PIG_CLASSPATH should do it
but it is not working on magellan.  so i added it to the jar file itself
at the root.  now it finds it.

also i've been adjusting the blast and velvet tests.  each work independently
but the output of blast does not work as input to velvet.  still working on it